+++ src/lib/libast/string/fmttv.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 
@@ -29,7 +30,7 @@ fmttv(const char* fmt, Tv_t* tv)
 {
 	char*	s;
 	char*	t;
-	int	n;
+	size_t	n;
 
 	s = fmttime(fmt, (time_t)tv->tv_sec);
 	if (!tv->tv_nsec || tv->tv_nsec == TV_NSEC_IGNORE)
